requirements.txt
Django==5.0.4
django-cors-headers==3.14.0
djangorestframework==3.14.0
python-dotenv
mysqlclient==2.2.4
django-cloudinary-storage==0.3.0
cloudinary==1.40.0